requirements.txt
aiogram>=3.3.0
SQLAlchemy>=2.0.0
greenlet>=3.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
alembic>=1.13.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
pytz>=2024.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
